requirements.txt
# MCP Server v2 Dependencies
mcp>=1.1.0
pyyaml>=6.0.0
